requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
supabase>=2.3.0
google-genai>=1.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
pdf2image>=1.16.0
Pillow>=10.0.0
